UIAndAudioupdatesPlusCleanup
-add exit to menu button to game over screen
-add exit to start screen on the racing scenes
-player car has been added to carmixer that is altered with car volume slider
-removed old waypoint system, coin collection system and many other scripts
