+++ smokin-tires/Assets/ClickGroup.mixer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 216106f6da55f9643a39041079cb66f3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 24100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
