--- smokin-tires/Assets/Scripts/DesertCheckpointsystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 187a5dcdb16c2d6479e80dc463859697
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
