--- smokin-tires/Assets/Scripts/DesertMap/FinishZone_desert.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b9fe489d8fccc934689a1764fb3c4f9b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
