--- smokin-tires/Assets/Scripts/DesertMap/WaypointDisablerDesert.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 02bdc11cf70a27c4b83124998dfc6d4b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
