--- smokin-tires/Assets/Scripts/LakeCountdown.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 64b1f19fae4633e45a84f18b05c2c5cd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
