+++ smokin-tires/Assets/Scripts/Lakeside/Finish_zone.cs
@@ -6,7 +6,7 @@ using UnityEngine.Audio;
 public class Finish_zone : MonoBehaviour
 {
 	//victory sound
-	public GameManager gameManager;
+	public LakeGameManager gameManager;
 
 	// When an object enters the finish zone, let the
 	// game manager know that the current game has ended
