--- smokin-tires/Assets/Scripts/Lakeside/GameManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0d34de4a38f10cf48811c5d6521f2be2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {fileID: 4162164826716764455, guid: 0000000000000000d000000000000000, type: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
