+++ smokin-tires/Assets/Scripts/Lakeside/LakeCountdown.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 64b1f19fae4633e45a84f18b05c2c5cd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
