+++ smokin-tires/Assets/Scripts/Lakeside/LakePlayerRespawnScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6586dcb9595991c488d18c4e10d959c4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
