--- smokin-tires/Assets/Scripts/SandMap/Car_respawn.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5796bc9929a47e64781975de19fd4e7b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
